requirements.txt
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
pyserial==3.5
setuptools==78.1.1
wheel==0.45.1
